requirements.txt
# Core framework
langchain==0.3.9
langchain-core==0.3.21
langchain-community==0.3.9
langchain-ollama
langchain-groq
langgraph==0.2.52

# Document processing
PyPDF2==3.0.1
pypdf==5.1.0
docx2txt==0.8
python-docx==1.1.2
pdfplumber==0.11.4

#image processing
python-jose==3.5.0
requests==2.32.4
# File type detection
filetype==1.2.0
python-magic==0.4.27

# Vector store and embeddings
faiss-cpu==1.9.0
numpy==1.26.4

# Data handling
pydantic==2.10.3

# Utilities
python-dotenv==1.0.1
typing-extensions==4.12.2
